Update java-helloworld.md
If you select/add Gradle through IntelliJ when creating a new project, I am not sure that `App.java` is neccessarily generated, like it is when you initialize gradle through command-line with `gradle init`. Therefore, I changed the sentence a bit to clarify for people who initialize project/gradle in the former way, that they may not see that specific file.