Update tutorial-subscription.md (#863)
`chargeCustomerForBillingPeriod` is called twice in the `for` loop, which would result in the customer getting charged twice. To achieve the intended behaviour of billing immediately when `BillingCycle` is called and then billing in regular intervals, we either need to bring out the first call outside of the `for` loop, or ger rid of the else block after `if (await wf.condition(...) {` entirely. I went for the first option in this PR.