Instead of nil activity, use method expression
Method expressions are nicer and get rid of the nil object.

They are not that widely known feature of go - they basically de-curry
a method and make it a function where the first parameter is the receiver -
https://go.dev/ref/spec#Method_expressions.

As Temporal SDK doesn't really care about the implementation and just looks at the method name
through runtime and reflection, this still works, and one gets rid of the "nil object" pattern.