Update the main CLI page with changes for 0.12
I manually tested all the steps for installing auto-completion for each
shell (bash, fish, zsh).

I have removed the "Install via cURL" options because it's generally bad
form, security-wise, to do anything that looks like `curl | sh`.  (Yes,
many people do it, however we don't want to encourage it.)

Note: I have not updated all the places with syntax changes in the
documentation generally; we'll need to do this in the future, but for
now, it's okay because the older syntax will continue to work.